requirements.txt
aiogram
requests
python-dotenv
PyPDF2